split sectionC part
